fix: add pages for deployments (Issue #1312) (#1313)
