+++ apps/ai-dial-admin/src/app/not-found.tsx
@@ -1,0 +1,5 @@
+import Page404 from '@/src/components/Page404/Page404';
+
+export default function NotFound() {
+  return <Page404 />;
+}
