+++ apps/ai-dial-admin/src/components/Page404/Page404.spec.tsx
@@ -7,12 +7,12 @@ import Page404 from './Page404';
 describe('Page404', () => {
   test('Should render not found page with icon, messages, and link', () => {
     render(<Page404 />);
-    expect(screen.getByText(ErrorI18nKey.PageNotFound)).toBeInTheDocument();
-    expect(screen.getByText(ErrorI18nKey.ResourceNotFound)).toBeInTheDocument();
-    expect(screen.getByText(ErrorI18nKey.CheckUrl)).toBeInTheDocument();
-    expect(screen.getByText(MenuI18nKey.Homepage)).toBeInTheDocument();
+    expect(screen.getByText('Page not found')).toBeInTheDocument();
+    expect(screen.getByText('Resource not found or no longer available.')).toBeInTheDocument();
+    expect(screen.getByText('Please check the URL or go back to the')).toBeInTheDocument();
+    expect(screen.getByText('Homepage')).toBeInTheDocument();
 
-    const link = screen.getByRole('link', { name: MenuI18nKey.Homepage });
+    const link = screen.getByText('Homepage');
     expect(link).toHaveAttribute('href', ApplicationRoute.Home);
   });
 });
